Move filters/mixins folder content with just index.js one level higher for better readability
