+++ core/filters/capitalize.js
