+++ core/filters/strip-html.js
