+++ core/mixins/onEscapePress.js
