+++ core/mixins/thumbnail.js
